bootsuccess: confirm reboot alt kernel
The btrfs image allows to install a second kernel. This patch allows to tell U-Boot booting
the alternate kernel was succesful.

Setting `fw_setenv switch 1` from linux will cause U-Boot to add 'New' to the kernel and initrd
name and set 'boot_count' to 1. If the subsequent boot succeeds 'boot_count' will be set to 0 in
linux by the bootsuccess service. This allows future boots to continue with the New kernel.

However if boot fails 'boot_count' remains 1. And on next reboot U-Boot the kernel will be
reset to the default one again.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>