image:btrfs: use zib compression and mixed while creating the image
It is cheaper to apply compression during image creation. While here
use mixed to remove the isolation and store data and metadatas in
the same block group type utilizes free space better and is recommended
for small devices < 5GiB which is the case for EMMC on Edison.
If you want to use an external SSD that is larger, you might want to remove
the "-M".

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>